+++ src/converter/latex/io.jl
@@ -207,46 +207,3 @@ function lx_literate(lxc::LxCom, lxd::Vector{LxDef})
     # then reprocess
     return reprocess(read(opath, String), lxd, nostripp=true)
 end
-
-#
-# UTILS
-#
-
-"""
-$SIGNATURES
-
-Internal function to process an array of strings to markdown table (in one
-single string). If header is empty, the first row of the file will be used as
-header.
-"""
-function csv2html(path::AS, header::AS)::String
-    csvcontent   = readdlm(path, ',', String, header=false)
-    nrows, ncols = size(csvcontent)
-    io = IOBuffer()
-    # writing the header
-    if ! isempty(header)
-        # header provided
-        newheader = split(header, ",")
-        hs = size(newheader,1)
-        if hs != ncols
-            return html_err("In `\\tableinput`: header size ($hs) and " *
-                            "number of columns ($ncols) do not match.")
-        end
-        write(io, prod("| " * h * " " for h in newheader))
-        rowrange = 1:nrows
-    else
-        # header from csv file
-        write(io, prod("| " * csvcontent[1, i] * " " for i in 1:ncols))
-        rowrange = 2:nrows
-    end
-    # writing end of header & header separator
-    write(io, "|\n|", repeat( " ----- |", ncols), "\n")
-    # writing content
-    for i in rowrange
-        for j in 1:ncols
-            write(io, "| ", csvcontent[i,j], " ")
-        end
-        write(io, "|\n")
-    end
-    return md2html(String(take!(io)))
-end
